Update registration pages.
