+++ websiteBackbone/templates/registration/edit_user_profile.html
@@ -1,6 +1,9 @@
 {% extends "websiteBackbone/header.html" %}
 {% block content %}
-<h2>Change password</h2>
+<div class="container-fluid">
+      <div class="row">
+        <div class="col-12 col-sm-6 col-lg-8">
+<h2>Change Password</h2>
 
 <form method="post">
     {% csrf_token %}
@@ -8,10 +11,14 @@
     <input type="submit" value="Change">
 </form>
 
+<h2>Update Account</h2>
+
 <form method="post">
   {% csrf_token %}
   {{ user_form.as_p }}
-  {{ profile_form.as_p }}
   <button type="submit">Save changes</button>
 </form>
+</div>
+</div>
+</div>
 {% endblock %}
