+++ websiteBackbone/templates/registration/password_change_done.html
@@ -8,8 +8,10 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <h2>Password changed</h2>
 
 <a href="{% url 'index' %}">Back to dashboard</a>
+</div>
 {% endblock %}
