+++ websiteBackbone/templates/registration/password_change_form.html
@@ -8,6 +8,7 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <h2>Change password</h2>
 
@@ -17,5 +18,14 @@
     <input type="submit" value="Change">
 </form>
 
+    <h2>Update Account</h2>
+
+<form method="post">
+  {% csrf_token %}
+  {{ user_form.as_p }}
+  <button type="submit">Save changes</button>
+</form>
+
 <a href="{% url 'index' %}">Back to dashboard</a>
+</div>
 {% endblock %}
