+++ websiteBackbone/templates/registration/password_reset_complete.html
@@ -8,8 +8,10 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <h2>Password reset complete</h2>
 
 <a href="{% url 'login' %}">Back to login</a>
+</div>
 {% endblock %}
