+++ websiteBackbone/templates/registration/password_reset_confirm.html
@@ -8,10 +8,12 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <form method="post">
     {% csrf_token %}
     {{ form.as_p }}
     <input type="submit" value="Confirm">
 </form>
+</div>
 {% endblock %}
