+++ websiteBackbone/templates/registration/password_reset_email.html
@@ -8,6 +8,7 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <h2>Login</h2>
 
@@ -15,4 +16,5 @@ Someone asked for password reset for email {{ email }}.
 Follow the link below:
 {{ protocol}}://{{ domain }}{% url 'password_reset_confirm' uidb64=uid token=token %}
 
+</div>
 {% endblock %}
