+++ websiteBackbone/templates/registration/password_reset_form.html
@@ -8,6 +8,7 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <h2>Send password reset link</h2>
 
@@ -18,4 +19,5 @@
 </form>
 
 <a href="{% url 'index' %}">Back to dashboard</a>
+</div>
 {% endblock %}
