+++ websiteBackbone/templates/registration/register.html
@@ -8,6 +8,7 @@
         display: none;
     }
 </style>
+    <div class="jumbotron jumbotron-fluid">
 
 <h2>Register</h2>
 
@@ -17,6 +18,7 @@
     <input type="submit" value="Register">
 </form>
 
-<a href="{% url 'index' %}">Back to dashboard</a>
+<a href="{% url 'index' %}">Back to dashboard</a> <br />
 <a href="{% url 'login' %}">Back to login</a>
+</div>
 {% endblock %}
