+++ websiteBackbone/urls.py
@@ -21,7 +21,6 @@ urlpatterns = [
     path(r'contact/', views.contact, name='contact'),
     path(r'submission/', views.submission, name='submission'),
     path(r'about/', views.about, name='about'),
-    path(r'register/', views.register, name="register"),
     path(r'privacypolicy/', views.privacypolicy, name='privacypolicy'),
     path(r'termsconditions/', views.termsconditions, name='termsconditions'),
     path(r'', views.index, name='index'),
